module-init-tools: 3.4 -> 3.16
svn path=/nixpkgs/branches/stdenv-updates/; revision=28621

+++ pkgs/build-support/kernel/modules-closure.sh
@@ -26,6 +26,7 @@ ensureDir $out/lib/modules/"$version"
 for module in $closure; do
     target=$(echo $module | sed "s^/nix/store/.*/lib/modules/^$out/lib/modules/^")
     if test -e "$target"; then continue; fi
+    if test \! -e "$module"; then continue; fi # XXX: to avoid error with "cp builtin builtin"
     mkdir -p $(dirname $target)
     echo $module
     cp $module $target
